+++ src/playground/src/components/playground/ClearButton.tsx
