+++ src/playground/src/components/playground/SliderComponent.tsx
