+++ src/playground/src/components/playground/SystemCard.tsx
